TOTS-91 Moved the characterize AOI function to the Additional Setup tab.
